+++ test/cases/optimization-imported-css-inline-source-map/expected/main.js
@@ -1,0 +1,19 @@
+/******/ (() => { // webpackBootstrap
+/******/ 	"use strict";
+var __webpack_exports__ = {};
+
+;// CONCATENATED MODULE: ./src/component.js
+
+
+console.log('>> component');
+
+;// CONCATENATED MODULE: ./src/main.js
+
+
+
+
+console.log('>> main');
+
+/******/ })()
+;
+//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoibWFpbi5qcyIsIm1hcHBpbmdzIjoiOzs7OztBQUF5Qjs7QUFFekI7OztBQ0ZxQjs7QUFFQTs7QUFFckIiLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9zcmMvY29tcG9uZW50LmpzIiwid2VicGFjazovLy8uL3NyYy9tYWluLmpzIl0sInNvdXJjZXNDb250ZW50IjpbImltcG9ydCAnLi9jb21wb25lbnQuY3NzJztcblxuY29uc29sZS5sb2coJz4+IGNvbXBvbmVudCcpO1xuIiwiaW1wb3J0ICcuL3N0eWxlLmNzcyc7XG5cbmltcG9ydCAnLi9jb21wb25lbnQnO1xuXG5jb25zb2xlLmxvZygnPj4gbWFpbicpO1xuIl0sIm5hbWVzIjpbXSwic291cmNlUm9vdCI6IiJ9