+++ test/cases/optimization-imported-css-inline-source-map/src/component.js
@@ -1,0 +1,3 @@
+import './component.css';
+
+console.log('>> component');