+++ test/cases/optimization-inline-imported-css-source-map/expected/main.js
@@ -1,0 +1,19 @@
+/******/ (() => { // webpackBootstrap
+/******/ 	"use strict";
+var __webpack_exports__ = {};
+
+;// CONCATENATED MODULE: ./src/component.js
+
+
+console.log('>> component');
+
+;// CONCATENATED MODULE: ./src/main.js
+
+
+
+
+console.log('>> main');
+
+/******/ })()
+;
+//# sourceMappingURL=main.js.map