+++ test/cases/optimization-inline-imported-css-source-map/src/main.js
@@ -1,0 +1,5 @@
+import './style.css';
+
+import './component';
+
+console.log('>> main');